+++ src/main/java/chess/domain/chesspiece/King.java
@@ -1,4 +1,4 @@
-package chess.domain.chessPiece;
+package chess.domain.chesspiece;
 
 import chess.domain.position.Direction;
 import chess.domain.position.Position;
@@ -7,15 +7,15 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import static chess.domain.chessPiece.Role.BLACK_BISHOP;
-import static chess.domain.chessPiece.Role.WHITE_BISHOP;
+import static chess.domain.chesspiece.Role.*;
 
-public class Bishop extends Piece{
+public class King extends Piece {
 
-    public Bishop(Team team) {
+    public King(Team team) {
         super(team);
     }
 
+    //TODO : 리팩토링
     @Override
     public List<Position> getRoute(Position source, Position target) {
         List<Position> route = new ArrayList<>();
@@ -35,16 +35,16 @@ public class Bishop extends Piece{
     protected void validateMovingRule(Position source, Position target) {
         int rowDistance = source.calculateRowDistance(target);
         int columnDistance = source.calculateColumnDistance(target);
-        if(rowDistance!=columnDistance){
+        if (rowDistance != 1 && columnDistance != 1) {
             throw new IllegalArgumentException("이동할 수 없습니다.");
         }
     }
 
     @Override
     public Role getRole() {
-        if(team.isWhite()) {
-            return WHITE_BISHOP;
+        if (team.isWhite()) {
+            return WHITE_KING;
         }
-        return BLACK_BISHOP;
+        return BLACK_KING;
     }
 }
