+++ src/main/java/chess/domain/chesspiece/Queen.java
@@ -1,16 +1,17 @@
-package chess.domain.chessPiece;
+package chess.domain.chesspiece;
 
+import chess.domain.position.Direction;
 import chess.domain.position.Position;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import static chess.domain.chessPiece.Role.*;
+import static chess.domain.chesspiece.Role.*;
 
-public class Knight extends Piece {
+public class Queen extends Piece {
 
-    public Knight(Team team) {
+    public Queen(Team team) {
         super(team);
     }
 
@@ -18,15 +19,22 @@ public class Knight extends Piece {
     public List<Position> getRoute(Position source, Position target) {
         List<Position> route = new ArrayList<>();
         validateMovingRule(source, target);
+
+        Direction direction = Direction.findDirection(source, target);
+        Position movingPosition = direction.move(source);
+
+        while (!movingPosition.equals(target)) {
+            route.add(movingPosition);
+            movingPosition = direction.move(movingPosition);
+        }
         return Collections.unmodifiableList(route);
     }
 
     @Override
     protected void validateMovingRule(Position source, Position target) {
         int rowDistance = source.calculateRowDistance(target);
-        int colDistance = source.calculateColumnDistance(target);
-
-        if (!(rowDistance == 2 && colDistance == 1 || rowDistance == 1 && colDistance == 2)) {
+        int columnDistance = source.calculateColumnDistance(target);
+        if (!source.isSameRow(target) && !source.isSameColumn(target) && rowDistance != columnDistance) {
             throw new IllegalArgumentException("이동할 수 없습니다.");
         }
     }
@@ -34,8 +42,8 @@ public class Knight extends Piece {
     @Override
     public Role getRole() {
         if (team.isWhite()) {
-            return WHITE_KNIGHT;
+            return WHITE_QUEEN;
         }
-        return BLACK_KNIGHT;
+        return BLACK_QUEEN;
     }
 }
